[#888] Change path for cache dir and simplify options
The data, sessions and sstore dirs were created in /etc/ckan/default.
This creates permission problems and is not the natural place to store
this kind of files. The main cache dir is now /tmp/s(ckan.site_id)s
(generally /tmp/default). The beaker dirs will be created there if not
defined in the config file. This will ease the packaging and simplify
the installation. Also our cron jobs for cleaning files look for them in
/tmp.