[#283] Save validated data dicts in activity_create
activity_create() was validating the data dict passed to it, but then
passing the original, unvalidated data dict on to model save. Make it pass
the validated data dict instead.

Fixing this reveals another problem, a JSONDecodeError when pulling
activities out of the database. Invalid JSON is getting into the data
column of the activity table. default_create_activity_schema() runs
unicode() on the 'data' field of the activity dict. This has the effect of
turning a Python dict (the 'data' field's value) into a Python string. The
schema should not do this, because the data field uses SQLAlchemy's
JsonDictType at the model level, so the dict is turned into a string at
the model level, it doesn't need to be turned into a string at the schema
level first.

To fix the second problem, just remove the unicode from
default_create_activity_schema.