[#1184] Update user checks in package create and update auth functions
The `auth_is_registered_user` function's name is misleading, as it only checks
if there is a user on the Pylons context object (ie if it is logged in).
It has been renamed to `auth_is_loggedin_user`, keeping the old as
deprecated. The function is not used anymore on the auth functions, as
the user should be always present in the context dict passed to the
functions (The controller sets context['user'] to c.user).