[#1199] Change 'Select an organization' to 'No organization'
When creating or editing a dataset, the dropdown for selecting the
dataset's organization contained an item labelled
'Select an organization...' that, if selected, would set the dataset to
belong to no organization (removing it from any organization it does
belong to). This seems quite confusing, change the label to
'No organization' to make it clear what it does.