[#1314] Single Solr schema file
To make easier to track changes and simplify the deplyment and upgrade,
the Solr schema file included in the CKAN source is now unified under
`schema.xml`.

As part of the release process, the file will be checked for changes,
and if present the version numbers on the schema file and the supported
versions list will be updated.

Release, install and upgrade docs have been updated to reflect this
changes.

Travis now also always points to schema.xml