Change DataStore ref in FileStore docs to seealso
It'll be good if we can use this style (a Sphinx .. seealso::, near the
top of the page) whenever we want to say "also see this page" in the
docs. That'll make it a lot more predictable and easier for users to
find, when one feature is connected to another.