[#1409] Fix the e-mail bodies' translation strings
We ended up creating a string with extra whitespace before the paragraphs,
like "...to be reset\n\n Please click the following...".

But the i18n message key doesn't have these whitespaces, so we always send
that in English. To solve it, we could either fix the message key in all .po
files, or change how we construct the string in mailer.py. I prefer the
latter.