[#1505] Fix bug when non-sysadmins invited users to organizations
The problem was with the .user_invite() auth function. Instead of adding 'id'
to the data_dict, I was adding it to the context. Fixed now.

I've also changed it to require a data_dict, and a data_dict['group_id'] (as
they're required anyway).

While I'm at it, I rewrote the tests into the new testing style.