[#1550] Show the file name when uploading a resource
Hiding it is bad makes it hard for the user to check its URL (although it makes
it more difficult for her to change it by mistake).

When you first upload, some browsers have a safety feature to not disclosing
the file's full path to the page so, instead of getting something like:

```
/home/vitor/datasets/unemployment.csv
```

You get:

```
C:\fakepath\unemployment.csv
```

To avoid bothering the user showing this ```fakepath``` stuff, we remove it,
showing just the file name. This only happens when uploading a file. The next
time you edit the resource, we'll have the entire URL, so that'll be shown
instead.