[#1386] Simplify docs for _get_action_user_name()
The documentation on how to create, for example, a Dataset with an specific
user is specific to the Dataset factory. The same can be said for any factory
that allows an user.

It would be good to use the same pattern (passing a user dict as "user" to the
factory) in all factories.