Update instructions to use containerised Solr/Pg by default
Getting CKAN to talk to Postgres and Solr on the Docker host is
perfectly possible, but explaining how to set up appropriate routes in
iptables is outside the scope of the CKAN installation docs.

This commit updates the CKAN run script to support connection to Pg and
Solr through Docker links, and updates the documentation accordingly.