[#1793] Don't delete all cookies whose names start with "ckan"
Apparently in 2012 it was necessary (possibly only on demo.ckan.org) to
delete all cookies whose names start with "ckan".

I don't think CKAN sets any cookies today, except for the session
cookie, so this seems unnecessary now.

Deleting all cookies whose names start with "ckan" is likely to delete
cookies created by extensions that are called something like "ckanext-foo",
which is very confusing, especially because the header that deletes the
cookies comes in the /en request not in the first request for the page.

Since I don't know the specific names of the cookies this code was
intended to delete, I'm just removing the code and guessing it's no
longer needed.